Fixed updater not updating
This was broken in d806829c11722656a4bd3ee7fd2f900b08aee04c when someone
removed an if checking if an error occured while downloading. It would
always assume the download failed and try to display a nonexistent error
causing a crash.

Fixed another semirelated issue where the download never actually started
because the code to start it was put inside the download progress update
event. Which never fired because it was never started.